+++ module/SecretaryBundle/Controller/RegistrationController.php
@@ -395,9 +395,9 @@ class RegistrationController extends \SecretaryBundle\Component\Controller\Regis
             ->getRepository('CommonBundle\Entity\General\Config')
             ->getConfigValue('secretary.isic_membership') == 1;
         $isicRedirect = false;
-        $isicAlreadyOrdered = $this->getEntityManager()
+        $isicOrder = $this->getEntityManager()
                         ->getRepository('CudiBundle\Entity\IsicCard')
-                        ->findByPersonAndYearQuery($person, $this->getCurrentAcademicYear())
+                        ->findByPersonAndYearQuery($academic, $this->getCurrentAcademicYear())
                         ->getResult();
 
         $membershipArticles = array();
@@ -418,15 +418,15 @@ class RegistrationController extends \SecretaryBundle\Component\Controller\Regis
         if (null !== $metaData) {
             if ($enableRegistration) {
                 if (null !== $metaData->getTshirtSize()) {
-                    $oldTshirtBooking = $this->getEntityManager()
-                        ->getRepository('CudiBundle\Entity\Sale\Booking')
-                        ->findOneAssignedByArticleAndPersonInAcademicYear(
-                            $this->getEntityManager()
-                                ->getRepository('CudiBundle\Entity\Sale\Article')
-                                ->findOneById($tshirts[$metaData->getTshirtSize()]),
-                            $academic,
-                            $this->getCurrentAcademicYear()
-                        );
+                    // $oldTshirtBooking = $this->getEntityManager()
+                    //     ->getRepository('CudiBundle\Entity\Sale\Booking')
+                    //     ->findOneAssignedByArticleAndPersonInAcademicYear(
+                    //         $this->getEntityManager()
+                    //             ->getRepository('CudiBundle\Entity\Sale\Article')
+                    //             ->findOneById($tshirts[$metaData->getTshirtSize()]),
+                    //         $academic,
+                    //         $this->getCurrentAcademicYear()
+                    //     );
                 }
             }
             $oldTshirtSize = $metaData->getTshirtSize();
@@ -512,7 +512,7 @@ class RegistrationController extends \SecretaryBundle\Component\Controller\Regis
                     }
 
                     if ($metaData->becomeMember() && null !== $selectedOrganization) {
-                        if ($isicMembership and !$isicAlreadyOrdered) {
+                        if ($isicMembership && $isicOrder == null) {
                             $isicRedirect = true;
                         } else {
                             $this->bookRegistrationArticles($academic, $organizationData['tshirt_size'], $selectedOrganization, $this->getCurrentAcademicYear());
@@ -527,7 +527,7 @@ class RegistrationController extends \SecretaryBundle\Component\Controller\Regis
                                     $this->getCurrentAcademicYear()
                                 );
 
-                            if (null !== $booking) {
+                            if (null !== $booking && $isicOrder->getBooking() !== $booking) {
                                 $this->getEntityManager()->remove($booking);
                             }
                         }
